backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "authlib>=1.6.7",
    "celery>=5.6.2",
    "fastapi>=0.128.5",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "logfire[fastapi,sqlalchemy,system-metrics]>=4.22.0",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "slowapi>=0.1.9",
    "sqladmin>=0.23.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "mkinit>=1.2.0",
]
